fix adc async so it actually works

+++ include/libhal-arm-mcu/rp/time.hpp
@@ -13,6 +13,6 @@ struct clock final : public hal::steady_clock
   u64 driver_uptime() override;
 };
 
-using microseconds = std::chrono::duration<u64, std::micro>;
+using microseconds = std::chrono::duration<u32, std::micro>;
 
 }  // namespace hal::rp::inline v4
